python/requirements.txt
fastapi
uvicorn
langchain-anthropic
langgraph
python-dotenv